+++ src/Ddd/ValueObject/Url.php
@@ -3,18 +3,18 @@
 declare(strict_types=1);
 
 /**
- * EmailAddress.php
+ * Url.php
  *
  * @author Michael Leßnau <michael.lessnau@gmail.com>
  * @since  2019-09-14
  */
 
-namespace D3\ValueObject;
+namespace Ddd\ValueObject;
 
 /**
- * Class EmailAddress
+ * Class Url
  */
-class EmailAddress extends SingleValueObject
+class Url extends SingleValueObject
 {
     /**
      * @param string $value Value.
@@ -24,9 +24,6 @@ class EmailAddress extends SingleValueObject
         $value
     ): bool {
         return is_string($value)
-            && filter_var(
-                $value,
-                FILTER_VALIDATE_EMAIL
-            ) !== false;
+            && filter_var($value, FILTER_VALIDATE_URL) !== false;
     }
 }