+++ tests/Ddd/ValueObject/EnumTest.php
@@ -9,9 +9,9 @@ declare(strict_types=1);
  * @since  2019-09-14
  */
 
-namespace Test\D3\ValueObject;
+namespace Test\Ddd\ValueObject;
 
-use D3\ValueObject\Enum;
+use Ddd\ValueObject\Enum;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 
@@ -25,8 +25,8 @@ class EnumTestDummy extends Enum
 /**
  * Class EnumTest
  *
- * @coversDefaultClass \D3\ValueObject\Enum
- * @covers \D3\ValueObject\SingleValueObject
+ * @coversDefaultClass \Ddd\ValueObject\Enum
+ * @covers \Ddd\ValueObject\SingleValueObject
  */
 class EnumTest extends TestCase
 {