+++ tests/Ddd/ValueObject/IpV4AddressTest.php
@@ -3,33 +3,33 @@
 declare(strict_types=1);
 
 /**
- * IpV6AddressTest.php
+ * IpV4AddressTest.php
  *
  * @author Michael Leßnau <michael.lessnau@gmail.com>
  * @since  2019-09-15
  */
 
-namespace Test\D3\ValueObject;
+namespace Test\Ddd\ValueObject;
 
-use D3\ValueObject\IpV6Address;
+use Ddd\ValueObject\IpV4Address;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 
 /**
- * Class IpV6AddressTest
+ * Class IpV4AddressTest
  *
- * @coversDefaultClass \D3\ValueObject\IpV6Address
- * @covers \D3\ValueObject\SingleValueObject
+ * @coversDefaultClass \Ddd\ValueObject\IpV4Address
+ * @covers \Ddd\ValueObject\SingleValueObject
  */
-class IpV6AddressTest extends TestCase
+class IpV4AddressTest extends TestCase
 {
     /**
      * @covers ::isValidValue
      *
      * @testWith [null, false]
      *           ["", false]
-     *           ["192.168.0.1", false]
-     *           ["2001:0db8:85a3:0000:0000:8a2e:0370:7334", true]
+     *           ["192.168.0.1", true]
+     *           ["2001:0db8:85a3:0000:0000:8a2e:0370:7334", false]
      *
      * @param mixed $value Value.
      * @param bool  $valid Indicates if the provided value is valid.
@@ -42,10 +42,10 @@ class IpV6AddressTest extends TestCase
             $this->expectExceptionMessage('Invalid value provided');
         }
 
-        $ipV6Address = new IpV6Address($value);
+        $ipV4Address = new IpV4Address($value);
 
         if ($valid) {
-            $this->assertSame($value, $ipV6Address->getValue());
+            $this->assertSame($value, $ipV4Address->getValue());
         }
     }
 
@@ -56,9 +56,9 @@ class IpV6AddressTest extends TestCase
      */
     public function testIsIpV4Address(): void
     {
-        $ipV6Address = new IpV6Address('2001:0db8:85a3:0000:0000:8a2e:0370:7334');
+        $ipV4Address = new IpV4Address('192.168.0.1');
 
-        $this->assertFalse($ipV6Address->isIpV4Address());
+        $this->assertTrue($ipV4Address->isIpV4Address());
     }
 
     /**
@@ -68,8 +68,8 @@ class IpV6AddressTest extends TestCase
      */
     public function testIsIpV6Address(): void
     {
-        $ipV6Address = new IpV6Address('2001:0db8:85a3:0000:0000:8a2e:0370:7334');
+        $ipV4Address = new IpV4Address('192.168.0.1');
 
-        $this->assertTrue($ipV6Address->isIpV6Address());
+        $this->assertFalse($ipV4Address->isIpV6Address());
     }
 }