+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v4.0.0]
+### Added
+- Unit tests
+### Modified
+- Root namespace is now "Ddd" (was "D3")
 ### Fixed
 - Documentation issue
+- Collection exception message
+- SingleValueObject::getFloatValue return type
 
 ## [v3.0.1] 2020-11-24
 ### Fixed